feat: introduce Sampler trait
