+++ crates/llm-base/src/lib.rs
@@ -14,7 +14,10 @@ mod quantize;
 mod vocabulary;
 
 pub mod model;
+pub mod samplers;
 pub mod util;
+
+use std::sync::Arc;
 
 pub use ggml;
 pub use ggml::Type as ElementType;
@@ -36,10 +39,11 @@ pub use model::{
 };
 pub use quantize::{quantize, QuantizeError, QuantizeProgress};
 pub use regex::Regex;
+pub use samplers::Sampler;
 pub use util::TokenUtf8Buffer;
 pub use vocabulary::{InvalidTokenBias, Prompt, TokenBias, TokenId, TokenizationError, Vocabulary};
 
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Debug)]
 /// The parameters for text generation.
 ///
 /// This needs to be provided during all inference calls,
@@ -50,20 +54,8 @@ pub struct InferenceParameters {
     /// Controls batch/chunk size for prompt ingestion in
     /// [InferenceSession::feed_prompt].
     pub n_batch: usize,
-    /// The top K words by score are kept during sampling.
-    pub top_k: usize,
-    /// The cumulative probability after which no more words are kept for sampling.
-    pub top_p: f32,
-    /// The penalty for repeating tokens. Higher values make the generation less
-    /// likely to get into a loop, but may harm results when repetitive outputs
-    /// are desired.
-    pub repeat_penalty: f32,
-    /// Temperature (randomness) used for sampling. A higher number is more random.
-    pub temperature: f32,
-    /// A list of tokens to bias against in the process of generation.
-    pub bias_tokens: TokenBias,
-    /// The number of tokens to consider for the repetition penalty.
-    pub repetition_penalty_last_n: usize,
+    /// The sampler to use for sampling tokens from the model's probabilities.
+    pub sampler: Arc<dyn Sampler>,
 }
 impl Default for InferenceParameters {
     /// Returns a reasonable default for the parameters.
@@ -74,12 +66,7 @@ impl Default for InferenceParameters {
         Self {
             n_threads: 8,
             n_batch: 8,
-            top_k: 40,
-            top_p: 0.95,
-            repeat_penalty: 1.30,
-            temperature: 0.80,
-            bias_tokens: TokenBias::empty(),
-            repetition_penalty_last_n: 512,
+            sampler: Arc::new(samplers::TopPTopK::default()),
         }
     }
 }
